+++ LuiExtended/modules/UnitFrames/_StaticControls/Group.lua
@@ -8,9 +8,7 @@ local FrameObject = LUIE_CustomFrameObject
 function FrameObject:UpdateGroupFrameStaticControls()
     FrameObject.UpdateStaticControlRoleIcon(self)
     FrameObject.UpdateStaticControlClassIcon(self)
-    if self.friendIcon then
-        self.friendIcon:SetHidden(true)
-    end
+    FrameObject.UpdateStaticControlFriendIcon(self)
     if FrameObject.HasCustomTopInfoFrameCategory(self) then
         FrameObject.UpdateTopInfoOverlandIcon(self)
         FrameObject.UpdateStaticControlNameLabel(self, "group")